Add dumpf() and dumps() methods. Old dump() method is an alias for dumpf().   dumpf() calls cleanup_old_timestamps() automatically. dumps() must call it   manually.
